3-4. 할 일 목록 부분을 위한 컴포넌트 생성하기(컴포넌트 분리하기)
- 할 일 목록 부분 컴포넌트로 분리
- 필요한 데이터 및 함수 props로 내려주기
