+++ src/App.js
@@ -1,32 +1,10 @@
 import React, {useState} from 'react';
 import './App.css';
+import List              from './components/List';
 
 export default function App() {
   const [todoData, setTodoData] = useState([]);
-  const [value, setValue] = useState("");
-
-  const btnStyle = {
-    color:        '#fff',
-    border:       'none',
-    padding:      '5px 9px',
-    borderRadius: '50%',
-    cursor:       'pointer',
-    float:        'right'
-  };
-
-  const getStyle = (completed) => {
-    return {
-      padding:        '10px',
-      borderBottom:   '1px #ccc dotted',
-      textDecoration: completed ? 'line-through' : 'none'
-    };
-  };
-
-  const handleClick = (id) => {
-    let newTodoData = todoData.filter((data) => data.id !== id);
-    console.log('new : ', newTodoData);
-    setTodoData(newTodoData);
-  };
+  const [value, setValue] = useState('');
 
   const handleChange = (e) => {
     console.log('e', e.target.value);
@@ -45,19 +23,7 @@ export default function App() {
 
     // 원래 있던 할 일에 새로운 할 일 추가주기
     setTodoData(prev => [...prev, newTodo]);
-    setValue("");
-  };
-
-  const handleCompleteChange = (id) => {
-    let newTodoData = todoData.map(data => {
-      if (data.id === id) {
-        data.completed = !data.completed;
-      }
-
-      return data;
-    });
-
-    setTodoData(newTodoData);
+    setValue('');
   };
 
   return (
@@ -67,15 +33,7 @@ export default function App() {
           <h1>할 일 목록</h1>
         </div>
 
-        {todoData.map(data => (
-          <div style={getStyle(data.completed)} key={data.id}>
-            <input type="checkbox"
-                   defaultChecked={false}
-                   onChange={() => handleCompleteChange(data.id)}/>
-            {data.title}
-            <button style={btnStyle} onClick={() => handleClick(data.id)}>x</button>
-          </div>
-        ))}
+        <List todoData={todoData} setTodoData={setTodoData}/>
 
         <form style={{display: 'flex'}} onSubmit={handleSubmit}>
           <input type="text"
